app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apify_fingerprint_datapoints==0.0.2
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
binaryornot==0.4.4
blessed==1.21.0
blinker==1.9.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
cookiecutter==2.6.0
Crawl4AI==0.6.3
cryptography==45.0.4
cssselect==1.3.0
curl_cffi==0.11.3
debugpy==1.8.11
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
easyocr==1.7.2
editor==1.6.6
email_validator==2.2.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
firecrawl==2.7.1
firecrawl-py==2.8.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.20.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
incremental==24.7.2
inquirer==3.4.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jaro-winkler==2.0.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy_loader==0.4
litellm==1.72.2
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
milvus-lite==2.4.12
more-itertools==10.7.0
motor==3.7.1
mpmath==1.3.0
multidict==6.4.4
narwhals==1.42.1
nest_asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numpy==2.3.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
openai==1.85.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parsel==1.10.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==25.1
platformdirs==4.3.8
playwright==1.52.0
prompt_toolkit==3.0.51
propcache==0.3.2
Protego==0.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.6.1
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.1
pymilvus==2.5.11
pymilvus.model==0.3.2
pymongo==4.13.0
PyMuPDF==1.26.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==3.17.4
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.8.0
rank-bm25==0.2.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
runs==1.2.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
Scrapy==2.13.2
sentence-transformers==4.1.0
service-identity==24.2.0
setuptools==78.1.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
stack_data==0.6.3
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.3.1
Twisted==25.5.0
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
w3lib==2.3.1
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
xmod==1.8.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
zstandard==0.23.0